Rename test_pytest as test_thztools and revise TestTHzGen for clarity
